refactor(routes): modularize auth routes and add master route mounting
